CB-13099 Improve environment update_load_balancers API response
Adds a new EnvironmentLoadBalancerUpdateResponse object that's returned with the update_load_balancer API is called.
The EnvironmentLoadBalancerService.updateLoadBalancerInEnvironmentAndStacks method now queries the flow service for
the most recent LoadBalancerUpdateFlowConfig flow, and includes that flow identifier in the response for tracking.

Tested by unit tests and running curl commands against the updated API.