CB-11740: Integrate with Telemetry solution.
Adding a new property to CDPEnvironmentAzureDetails, bool resourceEncryptionEnabled, this is set only when encryptionKeyUrl is present.
As encryptionKeyUrl is present only when "CDP_CB_AZURE_DISK_SSE_WITH_CMK" is enabled and encryptionKey is provided during environment creation, checking its presence is sufficient to know if user is using "Azure SSE with CMK".